service-inventaire/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
requests==2.31.0
drf-yasg==1.21.7